apply clippy error fixes
